+++ MachOView-master/capstone/arch/X86/X86IntelInstPrinter.c
@@ -395,7 +395,7 @@ static void printImm(MCInst *MI, SStream *O, int64_t imm, bool positive)
 					}
 				}
 
-				if (imm == 0x8000000000000000LL)  // imm == -imm
+				if (imm == (int64_t)0x8000000000000000LL)  // imm == -imm
 					SStream_concat0(O, "8000000000000000h");
 				else if (need_zero_prefix(imm))
 					SStream_concat(O, "0%"PRIx64"h", imm);
@@ -439,7 +439,7 @@ static void printImm(MCInst *MI, SStream *O, int64_t imm, bool positive)
 	} else {
 		if (MI->csh->syntax == CS_OPT_SYNTAX_MASM) {
 			if (imm < 0) {
-				if (imm == 0x8000000000000000LL)  // imm == -imm
+				if (imm == (int64_t)0x8000000000000000LL)  // imm == -imm
 					SStream_concat0(O, "8000000000000000h");
 				else if (imm < -HEX_THRESHOLD) {
 					if (need_zero_prefix(imm))
@@ -459,7 +459,7 @@ static void printImm(MCInst *MI, SStream *O, int64_t imm, bool positive)
 			}
 		} else {	// Intel syntax
 			if (imm < 0) {
-				if (imm == 0x8000000000000000LL)  // imm == -imm
+				if (imm == (int64_t)0x8000000000000000LL)  // imm == -imm
 					SStream_concat0(O, "0x8000000000000000");
 				else if (imm < -HEX_THRESHOLD)
 					SStream_concat(O, "-0x%"PRIx64, -imm);
